requirements.txt
absl-py
asgiref
astunparse
certifi
charset-normalizer
Django
django-restframework
flatbuffers
gast
google-pasta
grpcio
h5py
idna
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
namex
numpy
opt_einsum
optree
packaging
protobuf
psycopg
psycopg2
Pygments
requests
rich
six
sqlparse
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
tensorflow_intel
termcolor
typing_extensions
tzdata
urllib3
Werkzeug
whitenoise
wrapt
gunicorn
djangorestframework
Pillow
drf-extra-fields